backend/SEO-checker/py-services/requirements.txt
psycopg2-binary
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
playwright==1.40.0
beautifulsoup4==4.12.2
lxml
requests==2.31.0